Move read_config out of Autocorpus class
