+++ autocorpus/config.py
@@ -6,14 +6,34 @@ from enum import Enum
 from typing import Any
 
 
+def read_config(config_path: str) -> dict[str, Any]:
+    """Reads a configuration file and returns its content.
+
+    Args:
+        config_path: The path to the configuration file.
+
+    Returns:
+        dict: The content of the configuration file.
+
+    Raises:
+        FileNotFoundError: If the configuration file does not exist.
+        json.JSONDecodeError: If the configuration file is not a valid JSON.
+        KeyError: If the configuration file does not contain the expected "config" key.
+    """
+    with open(config_path, encoding="utf-8") as f:
+        ## TODO: validate config file here if possible
+        content = json.load(f)
+        return content["config"]
+
+
 class DefaultConfig(Enum):
     """An enumeration representing different configuration files for various datasets.
 
     Attributes:
-            LEGACY_PMC (str): Configuration file for legacy PMC data (pre-October 2024).
-            PMC (str): Configuration file for current PMC data.
-            PLOS_GENETICS (str): Configuration file for PLOS Genetics data.
-            NATURE_GENETICS (str): Configuration file for Nature Genetics data.
+            LEGACY_PMC: Configuration file for legacy PMC data (pre-October 2024).
+            PMC: Configuration file for current PMC data.
+            PLOS_GENETICS: Configuration file for PLOS Genetics data.
+            NATURE_GENETICS: Configuration file for Nature Genetics data.
 
     Methods:
             load_config():
@@ -27,14 +47,14 @@ class DefaultConfig(Enum):
     PLOS_GENETICS = "config_plos_genetics.json"
     NATURE_GENETICS = "config_nature_genetics.json"
 
-    def __init__(self, filename):
+    def __init__(self, filename: str) -> None:
         """Initializes the DefaultConfig enum with the given filename.
 
         Args:
-            filename (str): The name of the configuration file to load.
+            filename: The name of the configuration file to load.
         """
         self._filename = filename
-        self._config = None  # Lazy-loaded cache
+        self._config: dict[str, Any] = {}  # Lazy-loaded cache
 
     def load_config(self) -> dict[str, Any]:
         """Loads the configuration file when first accessed.
@@ -42,7 +62,7 @@ class DefaultConfig(Enum):
         Returns:
             The configuration file as a dictionary.
         """
-        if self._config is None:
+        if self._config == {}:
             config_path = resources.files("autocorpus.configs") / self._filename
             with config_path.open("r", encoding="utf-8") as f_in:
                 self._config = json.load(f_in)["config"]
